requirements.txt
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
